requirements.txt
python-dotenv==1.0.1
openai==1.12.0
requests==2.31.0
typing-extensions==4.9.0
python-dateutil==2.8.2 